third_party/blink/web_tests/external/wpt/css/css-color/contrast-color-currentcolor-inherited-expected.txt

This is a testharness.js-based test.
[FAIL] currentcolor inside contrast-color() computes to the keyword
  assert_equals: white is contrast to black expected "rgb(255, 255, 255)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
